Removed BloomMap.hpp and CountingBloomFilter.hpp. Renamed CountBloomFilter.hpp to CountingBloomFilter.hpp
